[example] Rename _example -> example due to dep
- if directory is put under `_example`, dep would ignore it saying not
go code when run `dep init`
- also there is no way let the example using a no vendored version of
libary, it won't build because `lib/vendor/lib2` is not same as
`example/vendor/lib` .... or what ever ....
