+++ example/github/.gitignore
