+++ example/github/Gopkg.toml
@@ -25,12 +25,36 @@
 #   unused-packages = true
 
 [[constraint]]
-  name = "github.com/go-sql-driver/mysql"
-  revision = "bc14601d1bd56421dd60f561e6052c9ed77f9daf"
+  branch = "master"
+  name = "github.com/dyweb/gommon"
+
+[[constraint]]
+  name = "github.com/google/go-github"
+  version = "15.0.0"
+
+[[constraint]]
+  name = "github.com/gorilla/sessions"
+  version = "1.1.0"
+
+[[constraint]]
+  name = "github.com/opentracing/opentracing-go"
+  version = "1.0.2"
 
 [[constraint]]
   name = "github.com/pkg/errors"
   version = "0.8.0"
+
+[[constraint]]
+  name = "github.com/spf13/cobra"
+  version = "0.0.1"
+
+[[constraint]]
+  name = "github.com/uber/jaeger-client-go"
+  version = "2.11.2"
+
+[[constraint]]
+  branch = "master"
+  name = "golang.org/x/oauth2"
 
 [prune]
   go-tests = true
