+++ example/github/Makefile
