+++ example/github/pkg/common/entity.go
