+++ example/github/pkg/common/version.go
