+++ example/github/pkg/server/auth/github.go
