+++ example/github/pkg/server/auth/pkg.go
@@ -1,7 +1,7 @@
 package auth
 
 import (
-	"github.com/at15/go.ice/_example/github/pkg/util/logutil"
+	"github.com/at15/go.ice/example/github/pkg/util/logutil"
 )
 
 var log = logutil.NewPackageLogger()
