+++ example/github/pkg/server/auth/register.go
