+++ example/github/pkg/server/pkg.go
@@ -2,7 +2,7 @@
 package server
 
 import (
-	"github.com/at15/go.ice/ice/util/logutil"
+	"github.com/at15/go.ice/example/github/pkg/util/logutil"
 )
 
 var log = logutil.NewPackageLogger()
