+++ example/github/pkg/service/auth.go
